services/clientes_finais_api/pyproject.toml
[tool.poetry]
name = "clientes-finais-api"
version = "0.1.0"
description = "API para o gerenciamento de Clientes Finais (CRUD)."
authors = ["Lucas Cid <lucas@vizubr.com>"]
readme = "README.md"
packages = [{include = "clientes_finais_api", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"

# --- Dependências internas do monorepo Vizu ---
vizu-observability-bootstrap = { path = "../../libs/vizu_observability_bootstrap" }
vizu-db-connector = { path = "../../libs/vizu_db_connector" }
vizu-shared-models = { path = "../../libs/vizu_shared_models" }


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
ruff = "^0.4.4"
sqlalchemy-utils = "^0.41.1" # <-- ADICIONE ESTA LINHA


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# --- Configuração Definitiva do Pytest para o Monorepo ---
[tool.pytest.ini_options]
pythonpath = [
    "src",
    "../../libs"
]